+++ NEWS.md
@@ -8,6 +8,7 @@
 **New features**
 
 * [#9](https://github.com/eblondel/d4storagehub4R/issues/9) Support `unshareItem` method
+* [#10](https://github.com/eblondel/d4storagehub4R/issues/10) Support `downloadItem` method
 
 ## [d4storagehub4R 0.3](https://cran.r-project.org/package=d4storagehub4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=d4storagehub4R)